moving calling backend request to details component
